% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTWL.R
\name{addTWL}
\alias{addTWL}
\title{Add sunrise sunset to Actogram}
\usage{
addTWL(date, offset, ...)
}
\arguments{
\item{date}{Date data in POSIXct format, most commonly \code{PAM_data$acceleration$date}}

\item{offset}{This parameter determines where the center of the graph is. When \code{offset = 0}, then midday is at the center of the graph. when \code{offset=12} midnight`}

\item{...}{Any additional parameters taken by graphics::points which the user may want to use to modify the graphic}
}
\value{
a plot
}
\description{
Add sunrise sunset to Actogram
}
