% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitudeCALC.R
\name{altitudeCALC}
\alias{altitudeCALC}
\title{Calculate altitude from pressure}
\usage{
altitudeCALC(P, T0 = 288.15, P0 = 1013.25)
}
\arguments{
\item{P}{on-bird pressure in hectopascals}

\item{T0}{temperature at sea level in kelvin (i.e. [celcius + 273.15] OR [(Farenheit − 32) * 5/9 + 273.15])}

\item{P0}{pressure at sea level in hectopascals}
}
\value{
altitude in metres, default is calculated according to International Standard Atmosphere model (International Organization for Standardization 1975: ISO 2533:1975)
}
\description{
This opens a java application which allows the user to zoom in and out. In Rstudio it will open in the viewer pane and in base R in an html readers. Note that this can be a bit slow
}
\examples{
#load dummy data
data(hoopoe)
PAM_data=hoopoe

altitude = altitudeCALC(P = PAM_data$pressure$obs)
plot(PAM_data$pressure$date, altitude, type="o",pch=16,
     xlab="Date", ylab="Altitude (m)")

}
