% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clockdrift.R
\name{calculate_clockdrift}
\alias{calculate_clockdrift}
\title{Clock Drift Adjustment}
\usage{
calculate_clockdrift(time, start, end)
}
\arguments{
\item{time}{a vector of POSIXct times.}

\item{start}{new start time as POSIXct.}

\item{end}{new end time as POSIXct.}
}
\description{
Adjust time for clock drift
}
\details{
Linearly rescale a sequence of dates to a new start and end time
to correct for clock drift in the tag.
}
\examples{
dates = hoopoe$magnetic$date
drift = 12
adjusted_end = as.POSIXct(dplyr::last(hoopoe$magnetic$date) + 12*60, tz="UTC")
drift_corrected = calculate_clockdrift(time = dates,
                             start= dates[1],
                             end = adjusted_end)

}
