% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_depth.R
\name{calculate_depth}
\alias{calculate_depth}
\title{Calculate depth from pressure}
\usage{
calculate_depth(P, p = 1023.6, g = 9.80665)
}
\arguments{
\item{P}{pressure in hectopascals}

\item{p}{water density for seawater is 1023.6kg/m3 and freshwater 997.0474 kg/m3}

\item{g}{gravitational field strength default 9.80665 m/s2}
}
\value{
depth in m
}
\description{
This function calculates depth from pressure
}
\examples{
depth = calculate_depth(11065)

}
