% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_triaxial_acceleration.R
\name{calculate_triaxial_acceleration}
\alias{calculate_triaxial_acceleration}
\title{Tri-axial acceleration functions}
\usage{
calculate_triaxial_acceleration(dta)
}
\arguments{
\item{dta}{magentic data from PAM logger see hoopoe$magnetic for an example}
}
\value{
roll, pitch and yaw from acceleration data
}
\description{
This function calculates roll, pitch and yaw from tri-axial acceleration data. It also centeres the acceleration data (for plotting on a sphere).
}
\examples{
data(swift)
PAM_data = swift

calculate_triaxial_acceleration(dta = PAM_data$magnetic)

}
\references{
Bidder, O.R., Walker, J.S., Jones, M.W., Holton, M.D., Urge, P., Scantlebury, D.M., Marks, N.J., Magowan, E.A., Maguire, I.E. and Wilson, R.P., 2015. Step by step: reconstruction of terrestrial animal movement paths by dead-reckoning. Movement ecology, 3(1), p.23.
}
