% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_water_pressure.R
\name{calculate_water_pressure}
\alias{calculate_water_pressure}
\title{Calculate water pressure from depth}
\usage{
calculate_water_pressure(h, p = 1023.6, g = 9.80665)
}
\arguments{
\item{h}{depth in m *must be positive}

\item{p}{water density for seawater is 1023.6kg/m3 and freshwater 997.0474 kg/m3}

\item{g}{gravitational field strength default 9.80665 m/s2}
}
\value{
pressure in hectopascals
}
\description{
This function calculates water pressure from depth
}
\examples{
pressure = calculate_water_pressure(113.16)

}
