% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_flap.R
\name{classify_flap}
\alias{classify_flap}
\title{Classify flapping flight}
\usage{
classify_flap(dta, period = 12, to_plot = TRUE, method = "kmeans", tz = "UTC")
}
\arguments{
\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 3x5min = 15 minutes of intense activity is considered flight}

\item{to_plot}{can be true or false. If true then threshold is plotted according to plot_histogram()}

\item{method}{for the time being only supports "kmeans", but will later also include maybe}

\item{tz}{timezone, default is "UTC"}
}
\value{
timetable: a timetable for when the species was migrating or not,

classification: a classification timeseries where datetime corresponds to activity, and

no_movement: the value in classification which corresponds to no movement

low_movement: the value in classification which corresponds to low activity

high_movement: the value in classification which corresponds to high activity

migration: the value in classification which corresponds to migratory flapping flight

threshold: the threshold between low and high activity
}
\description{
This function uses activity data to classify migratory flapping flight.
}
\examples{
#specify the data location
data(hoopoe)
# make sure the cropping period is in the correct date format
start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")

# Crop the data
PAM_data= create_crop(hoopoe,start,end)
str(PAM_data)

behaviour = classify_flap(dta = PAM_data$acceleration, period = 12)

col=c("brown","cyan4","black","gold")
plot(PAM_data$acceleration$date[6000:8000],PAM_data$acceleration$act[6000:8000],
     col=col[behaviour$classification][6000:8000],
     type="o", pch=20,
     xlab="Date",
     ylab="Activity")

behaviour$timetable

}
\references{
Bäckman, J., Andersson, A., Alerstam, T., Pedersen, L., Sjöberg, S., Thorup, K. and Tøttrup, A.P., 2017. Activity and migratory flights of individual free‐flying songbirds throughout the annual cycle: method and first case study. Journal of avian biology, 48(2), pp.309-319.

Liechti, F., Bauer, S., Dhanjal-Adams, K.L., Emmenegger, T., Zehtindjiev, P. and Hahn, S., 2018. Miniaturized multi-sensor loggers provide new insight into year-round flight behaviour of small trans-Sahara avian migrants. Movement ecology, 6(1), p.19.

Bruderer, B., Peter, D., Boldt, A. and Liechti, F., 2010. Wing‐beat characteristics of birds recorded with tracking radar and cine camera. Ibis, 152(2), pp.272-291.
}
