% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_flapping.R
\name{classify_flapping}
\alias{classify_flapping}
\title{Make timetable}
\usage{
classify_flapping(dta, flapping_duration = 3)
}
\arguments{
\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{flapping_duration}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 3x5min = 15 minutes of intense activity is considered a migratory flight}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
#specify the data location
data(PAM_data)
str(PAM_data)

#plot the activity to see if it looks ok
plot(PAM_data$acceleration$date, PAM_data$acceleration$act, xlab="Time", ylab="activity")

# at first glance it looks like the logger was removed off a birds and left in arucksack
PAM_data$acceleration = PAM_data$acceleration[((PAM_data$acceleration$date >= "2016-07-30")
& (PAM_data$acceleration$date <= "2017-06-01")),]

behaviour = classify_flapping(dta = PAM_data$acceleration, flapping_duration = 3)


col=col=c("brown","cyan4","black","gold")
plot(PAM_data$acceleration$date[2000:4000],PAM_data$acceleration$act[2000:4000],
col=col[behaviour$classification][2000:4000], type="o", pch=20, xlab="Date", ylab="Activity")

behaviour$timetable

}
