% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_pressurechange.R
\name{classify_pressurechange}
\alias{classify_pressurechange}
\title{Classify pressure change}
\usage{
classify_pressurechange(dta, thld = 2, duration = 1, tz = "UTC")
}
\arguments{
\item{dta}{data stored as a list see str(hoopoe) for example format}

\item{thld}{threshold that is used di discinguish between background pressure fluctuations and flights or dives, by default it is 2hPa}

\item{duration}{duration in hours of the pressure displacement used to quantify a long duration dive or flight, default is 1h}

\item{tz}{timezone, default is "UTC"}
}
\value{
timetable: a timetable for when the species was migrating or not,

classification: a classification timeseries where datetime corresponds to activity, and

no_pressurechange: the value in classification which corresponds to background pressure changes from weather

short_pressurechange: the value in classification which corresponds to a short change in pressure

long_pressurechange: the value in classification which corresponds to a long change in pressure

threshold: the threshold between low and high activity
}
\description{
This function uses a change in pressure greater than a certain threshold to classify diving or flying
}
\examples{
#specify the data location
data(hoopoe)
# make sure the cropping period is in the correct date format
start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")

# Crop the data
PAM_data= create_crop(hoopoe,start,end)

behaviour = classify_pressurechange(dta = PAM_data$pressure)

col=c("black","cyan4","gold")
plot(PAM_data$pressure$date[2000:2800],PAM_data$pressure$obs[2000:2800],
     col=col[behaviour$classification+1][2000:2800],
     type="o", pch=20,
     xlab="Date",
     ylab="Pressure")

behaviour$timetable

}
\references{
Bäckman, J., Andersson, A., Alerstam, T., Pedersen, L., Sjöberg, S., Thorup, K. and Tøttrup, A.P., 2017. Activity and migratory flights of individual free‐flying songbirds throughout the annual cycle: method and first case study. Journal of avian biology, 48(2), pp.309-319.

Liechti, F., Bauer, S., Dhanjal-Adams, K.L., Emmenegger, T., Zehtindjiev, P. and Hahn, S., 2018. Miniaturized multi-sensor loggers provide new insight into year-round flight behaviour of small trans-Sahara avian migrants. Movement ecology, 6(1), p.19.

Bruderer, B., Peter, D., Boldt, A. and Liechti, F., 2010. Wing‐beat characteristics of birds recorded with tracking radar and cine camera. Ibis, 152(2), pp.272-291.
}
