% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyFLAP.R
\name{classifyFLAP}
\alias{classifyFLAP}
\title{Make timetable}
\usage{
classifyFLAP(dta, period = 12, toPLOT = TRUE, method = "kmeans",
  tz = "UTC")
}
\arguments{
\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 3x5min = 15 minutes of intense activity is considered flight}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{method}{for the time being only supports "kmeans", but will later also include maybe}

\item{tz}{timezone, default is "UTC"}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
#specify the data location
data(hoopoe)
# make sure the cropping period is in the correct date format
start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")

# Crop the data
PAM_data= cutPAM(hoopoe,start,end)
str(PAM_data)

behaviour = classifyFLAP(dta = PAM_data$acceleration, period = 12)

col=c("brown","cyan4","black","gold")
plot(PAM_data$acceleration$date[6000:8000],PAM_data$acceleration$act[6000:8000],
     col=col[behaviour$classification][6000:8000],
     type="o", pch=20,
     xlab="Date",
     ylab="Activity")

behaviour$timetable

}
