% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyLIGHT.R
\name{classifyLIGHT}
\alias{classifyLIGHT}
\title{Make timetable}
\usage{
classifyLIGHT(dta, period = 3, toPLOT = TRUE, light_threshold = 5,
  duration_threshold = 2, keep_one_off_missclassifications = TRUE,
  method = c("manual", "kmeans"), tz = "UTC")
}
\arguments{
\item{dta}{the input light data to be classified e.g. bee_eater$light}

\item{period}{number of timepoints after which behaviour is considered soaring e.g. for light recorded every 5 minutesand period = 3, then 3x5min = 15 minutes of continuous light is considered soaring flight}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{light_threshold}{This is the threshold between nightime and daystime}

\item{duration_threshold}{the duration in hours where flight is not longer soaring, but migratory}

\item{keep_one_off_missclassifications}{Can be TRUE or FALSE, if true, any one-off soaring missclassifications are replaced with soaring}

\item{method}{for the time being only supports "kmeans" and "manual", if methods= "kmeans" then duration_threshold is overwritten}

\item{tz}{timezone, default is "UTC"}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
data("bee_eater")

start = as.POSIXct("2015-07-12","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2016-06-01","\%Y-\%m-\%d", tz="UTC")

bee_eater = cutPAM(bee_eater, start, end)
quickPLOT(bee_eater, measurements = c("pressure", "light"))

dta = bee_eater$light

behaviour = classifyLIGHT(dta , period = 3, toPLOT = TRUE,
light_threshold = 5, #duration_threshold = 2,
keep_one_off_missclassifications = TRUE,
method = "kmeans", tz= "UTC")


col=c("brown","cyan4","black","gold")
plot(dta$date,dta$obs, type="o", col= col[behaviour$classification], pch=16,
cex=ifelse(behaviour$classification == behaviour$migrating,1,0))

behaviour$timetable

}
