% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMIG.R
\name{classifyMIG}
\alias{classifyMIG}
\title{Make timetable}
\usage{
classifyMIG(pressure = PAM_data$pressure, light = PAM_data$light,
  threshold = 1, tz = "UTC")
}
\arguments{
\item{tz}{timezone, default is "UTC"}

\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 3x5min = 15 minutes of intense activity is considered flight}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{method}{for the time being only supports "kmeans", but will later also include maybe}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
#specify the data location
data(PAM_data)
str(PAM_data)


#First start by finding the threshold between night and day
plot(PAM_data$light$date[2000:4000],PAM_data$light$obs[2000:4000], type="l")

#plot the activity to see if it looks ok
plot(PAM_data$acceleration$date, PAM_data$acceleration$act, xlab="Time", ylab="activity")

# at first glance it looks like the logger was removed off a birds and left in arucksack
PAM_data$acceleration = PAM_data$acceleration[((PAM_data$acceleration$date >= "2016-07-30")
& (PAM_data$acceleration$date <= "2017-06-01")),]

behaviour = classifyFLAP(dta = PAM_data$acceleration, period = 4)


col=col=c("brown","cyan4","black","gold")
plot(PAM_data$acceleration$date[2000:4000],PAM_data$acceleration$act[2000:4000],
col=col[behaviour$classification][2000:4000], type="o", pch=20, xlab="Date", ylab="Activity")

behaviour$timetable

}
