% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climbPWR.R
\name{climbPWR}
\alias{climbPWR}
\title{Derive classification data for soaring birds}
\usage{
climbPWR(mass, acceleration, height, duration)
}
\arguments{
\item{mass}{in kg}

\item{acceleration}{in m/s}

\item{height}{in m}

\item{duration}{in seconds}
}
\value{
climb power based on: Van Walsum, T. A., Perna, A. , Bishop, C. M., Murn, C. P., Collins,
P. M., Wilson, R. P. and Halsey, L. G. (2019), Exploring the relation between flapping behaviour and accelerometer signal during ascending flight,
and a new approach to calibration. Ibis. doi:10.1111/ibi.12710
}
\description{
Derive classification data for soaring birds
}
\examples{
data(bee_eater)
PAM_data=bee_eater
twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs,
                             LightThreshold = 2, ask = FALSE)

TOclassify = soarPREP(dta = PAM_data,
                      twl = twl)

power = climbPWR(mass = 0.057, #in kg
                 acceleration = TOclassify$sum_activity / (TOclassify$duration * 3600) , # in m/s
                 height = TOclassify$cum_altitude_up, # in m
                 duration = TOclassify$duration * 3600 # in seconds
)


}
