% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMAT.R
\name{confusionMAT}
\alias{confusionMAT}
\title{Calculate error and create a confusion matrix}
\usage{
confusionMAT(reference, classified)
}
\arguments{
\item{reference}{Reference dataset wich classification is to be compared to}

\item{classified}{Classification output to compare with a reference dataset}
}
\value{
a confusion matrix
}
\description{
Calculate error and create a confusion matrix
}
\examples{

# Get hoopoe data and crop where appropriate
data("hoopoe")
start = as.POSIXct("2016-08-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-05-15","\%Y-\%m-\%d", tz="UTC")
PAM_data = cutPAM(hoopoe, start, end)

# Generate a classification from activity
activity_classification = classifyFLAP(PAM_data$acceleration, period = 10)
reference = activity_classification$classification
reference[reference != activity_classification$migration] = "Not Migrating"
reference[(activity_classification$classification ==
          activity_classification$high_movement)] = "Flying"
reference[reference == activity_classification$migration] = "Migrating"

# Generate a very simple pressure classification based on pressure changes
classified  = ifelse( abs(diff(PAM_data$pressure$obs)) >2, "Migrating", "Not Migrating")

# Remove dates from activity which are not also present in pressure
to_keep = which(PAM_data$acceleration$date \%in\% PAM_data$pressure$date)
reference = reference[to_keep]
reference = reference[-1]

confusionMAT(reference, classified)

}
