% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_diveMove.R
\name{convert_diveMove}
\alias{convert_diveMove}
\title{convert diveMove input into pamlr input}
\usage{
convert_diveMove(
  input,
  id = 0,
  measurements = c("light", "depth", "temperature")
)
}
\arguments{
\item{input}{diveMove data}

\item{id}{individual identifyer for tracked animal e.g. leg tag number}

\item{measurements}{a series of measurements logged by the dtr logger which are to be imported. Currently supports: "light", "depth","temperature"}
}
\value{
a list of measurements for the one individual
}
\description{
Converts  diveMove input into pamlr input
}
\examples{
#install.packages("diveMove")
#library(diveMove)
#zz <- system.file(file.path("data", "dives.csv"),
#                  package="diveMove", mustWork=TRUE)
#(sealX <- readTDR(zz, speed=TRUE, sep=";", na.strings="", as.is=TRUE))

#PAM_data = convert_diveMove(sealX, id="sealX")

}
