% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphClassified.R
\name{dygraphClassified}
\alias{dygraphClassified}
\title{Plot PAM data with dygraphs}
\usage{
dygraphClassified(dta, from = dta$light$date[1],
  to = dta$light$date[length(dta$light$date)], toPLOT = names(dta),
  timetable = timetable)
}
\arguments{
\item{dta}{PAM data to be plotted}

\item{from}{date that plotting starts}

\item{to}{date that plotting ends}

\item{toPLOT}{names of the variables to plot. For now this includes `light`, `pressure`, `acceleration` and `temperature`}

\item{timetable}{classification of start/stop}
}
\value{
a plot of all the measurements
}
\description{
This opens a java application which allows the user to zoom in and out. In Rstudio it will open in the viewer pane and in base R in an html readers. Note that this can be a bit slow
}
\examples{
#load dummy data
data(PAM_data)

# This bit is for Rstudio users to prevent html from opening in Viewer pane and crashing
# It opens in web browser instead
backup_options <- options()
options(viewer=NULL)

# Classify bir's behaviour
behaviour = classifyFLAP(dta = PAM_data$acceleration, period = 5)

# Plot interactive graphics
dygraphClassified(dta = PAM_data, timetable = behaviour$timetable)

# restore Rstudio settings from before plot
options(backup_options)

}
