% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphPAM.R
\name{dygraphPAM}
\alias{dygraphPAM}
\title{Plot PAM data}
\usage{
dygraphPAM(dta, from = dta$light$date[1],
  to = dta$light$date[length(dta$light$date)], toPLOT = names(dta))
}
\arguments{
\item{dta}{PAM data to be plotted}
}
\value{
a plot of all the measurements
}
\description{
Plot PAM data
}
\examples{
#specify the data location
data(PAM_data)
str(PAM_data)

#plot the activity to see if it looks ok
plot(PAM_data$acceleration$date, PAM_data$acceleration$act, xlab="Time", ylab="activity")

# at first glance it looks like the logger was removed off a birds and left in arucksack
PAM_data$acceleration = PAM_data$acceleration[((PAM_data$acceleration$date >= "2016-07-30")
& (PAM_data$acceleration$date <= "2017-06-01")),]


}
