% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hoopoe.R
\docType{data}
\name{hoopoe}
\alias{hoopoe}
\title{PAM data collected from Hoopoe}
\format{List of 6 variables
\describe{
  \item{id}{Logger id}
  \item{pressure}{Date in UTC and Atmospheric pressure measurements in hectopascals (every 15 minutes)}
  \item{light}{Date in UTC and Light intensity (every 5 minutes)}
  \item{acceleration}{Date in UTC, pitch of the accelerometer (i.e. an angle), and derived activity (every 5 minutes)}
  \item{temperature}{Date in UTC and temperature on bird in celcius(every 15 minutes)}
  \item{magnetic}{Date in UTC and magnetic measurements on 3 axes(every 15 minutes)}
}}
\source{
\url{www.vogelwarte.ch/en}
}
\usage{
hoopoe
}
\description{
A dataset containint pressure, light, acceleration, temperature and magnetism from a hoopoe (id = 16AJ)
}
\keyword{datasets}
