% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_PAM.R
\name{Import_PAM}
\alias{Import_PAM}
\title{Import PAM data}
\usage{
Import_PAM(pathname, measurements = c(".pressure", ".glf", ".acceleration",
  ".temperature", ".magnetic"))
}
\arguments{
\item{pathname}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be imported. Currently supports these file extentions: .pressure", ".glf", ".gle",".acceleration", ".temperature" and ".magnetic"}
}
\value{
a list of measurements for the individual measured
}
\description{
Import PAM data
}
\examples{
pathname = "R:/40 Data/20 Geolocator/10 Raw data/UpuEpoCH16/16AJ_20170809"
measurements = c(".pressure", ".glf")
PAM_data = Import_PAM(pathname, measurements)
str(PAM_data)


}
