% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importPAM.R
\name{importPAM}
\alias{importPAM}
\title{Import PAM data}
\usage{
importPAM(pathname = pathname, measurements = c(".pressure", ".glf",
  ".acceleration", ".temperature", ".magnetic"))
}
\arguments{
\item{pathname}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be imported. Currently supports these file extentions: ".pressure", ".glf", ".gle",".acceleration", ".temperature", "AirTemperature" and ".magnetic"}
}
\value{
a list of measurements for the one individual
}
\description{
Imports and formats many datasets into one big nested list containing all the data from the different sensors. A subset of sensors can be selected using `measurements`.
}
\examples{
pathname = "R:/40 Data/20 Geolocator/10 Raw data/UpuEpoCH16/16AJ_20170809"
measurements = c(".pressure", ".glf")
PAM_data = importPAM(pathname, measurements)
str(PAM_data)
plot(PAM_data$light$date[3000:5000], PAM_data$light$obs[3000:5000],
type="l", xlab="Date", ylab="Light Intensity")

}
