% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magSPHERE.R
\name{magSPHERE}
\alias{magSPHERE}
\title{Plot calibrated magnetic data on a sphere}
\usage{
magSPHERE(magX, magY, magZ, col = "black", ...)
}
\arguments{
\item{magX}{magentic data on x axis}

\item{magY}{magentic data on y axis}

\item{magZ}{magentic data on z axis}

\item{col}{color of points}

\item{...}{additional input for rgl::spheres3d for plotting points on the sphere using}
}
\value{
a 3d sphere plot
}
\description{
Plot calibrated magnetic data on a sphere
}
\examples{
data("swift")
start = as.POSIXct("2016-09-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-04-15","\%Y-\%m-\%d", tz="UTC")
swift = cutPAM(swift, start, end)
PAM_data = swift

calibration = triMAG(dta = PAM_data$magnetic)

# magSPHERE(magX = calibration$calib_magx,
#           magY = calibration$calib_magy,
#           magZ = calibration$calib_magz)

}
