% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePAM.R
\name{mergePAM}
\alias{mergePAM}
\title{merges all the data into one table for interpolation}
\usage{
mergePAM(dta, interp = TRUE, keep_NA = TRUE)
}
\arguments{
\item{dta}{raw pam data see `data(bee_eater` for example}

\item{interp}{logical. whether or not to interpolate if there are NAs}

\item{keep_NA}{logical. Whether or not to keep NA}
}
\value{
merged and interpolated dataset
}
\description{
merges all the data into one table for interpolation
}
\examples{
data(bee_eater)
PAM_data = bee_eater
merged_dta = mergePAM(PAM_data, interp = TRUE)
head(merged_dta)

}
