% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSOARFLAP.R
\name{mergeSOARFLAP}
\alias{mergeSOARFLAP}
\title{Compare two different classifications}
\usage{
mergeSOARFLAP(dta = PAM_data, interpolate = F)
}
\arguments{
\item{dta}{data stored as a list. see data(PAM_data) then str(PAM_data) for example format}

\item{interpolate}{can be either TURE or FALSE. Use with caution. If TRUE, then linearly interpolates pressure observations}
}
\value{
merged observations and classifications
}
\description{
Compare two different classifications
}
\examples{
#specify the data location
data(PAM_data)
str(PAM_data)


# at first glance it looks like the logger was removed off a birds and left in arucksack
#  # so remove un-needed data
PAM_data$acceleration = PAM_data$acceleration[((PAM_data$acceleration$date >= "2016-07-30")
                                               & (PAM_data$acceleration$date <= "2017-06-01")),]

PAM_data$pressure = PAM_data$pressure[((PAM_data$acceleration$date >= "2016-07-30")
                                               & (PAM_data$acceleration$date <= "2017-06-01")),]

mergeTBL = mergeSOARFLAP(dta = PAM_data, interpolate = F)
plot(mergeTBL$date[!is.na(mergeTBL$obs)][500:1500],mergeTBL$obs[!is.na(mergeTBL$obs)][500:1500],
           type="o", pch=16,
           col=col[mergeTBL$classSOAR[!is.na(mergeTBL$obs)][500:1500]])
plot(mergeTBL$date[!is.na(mergeTBL$obs)][500:1500],mergeTBL$obs[!is.na(mergeTBL$obs)][500:1500],
             type="o", pch=16,
             col=col[mergeTBL$classFLAP[!is.na(mergeTBL$obs)][500:1500]])

}
