% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pam3D.R
\name{pam3D}
\alias{pam3D}
\title{3d scatterplot}
\usage{
pam3D(x, y, z, ...)
}
\arguments{
\item{x}{data to plot on x axis}

\item{y}{data to plot on y axis}

\item{z}{data to plot on z axis}

\item{...}{any additional parameters used by rgl::plot3d}
}
\value{
a 3d scatter plot
}
\description{
Creates an interactive 3d scatterplot
}
\examples{
#data("swift")
#start = as.POSIXct("2016-09-01","\%Y-\%m-\%d", tz="UTC")
#end = as.POSIXct("2017-04-15","\%Y-\%m-\%d", tz="UTC")
#swift = cutPAM(swift, start, end)
#PAM_data = swift

#calibration = triMAG(dta = PAM_data$magnetic)

#pam3D(PAM_data$magnetic$mX, PAM_data$magnetic$mY, PAM_data$magnetic$mZ,
#       xlab= "X", ylab= "Y", zlab= "Z")

}
\references{
Adler, D., Nenadic, O. and Zucchini, W., 2003, March. Rgl: A r-library for 3d visualization with opengl. In Proceedings of the 35th Symposium of the Interface: Computing Science and Statistics, Salt Lake City (Vol. 35).
}
