% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMLr.R
\name{pamlr}
\alias{pamlr}
\title{pamlr}
\description{
This package manipulates data from SOI-GDL3pam loggers (developped by the Swiss Ornithological Institute). These measure Pressure, Activity, Magnetism and Light.
}
\author{
Kiran Dhanjal-Adams \email{kiran.dhanjaladams@uqconnect.edu.au}
}
\keyword{"_PACKAGE"}
\keyword{internal}
