% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3D.R
\name{plot3D}
\alias{plot3D}
\title{3d Plot}
\usage{
plot3D(x, y, z, ...)
}
\arguments{
\item{x}{data to plot on x axis}

\item{y}{data to plot on y axis}

\item{z}{data to plot on z axis}

\item{...}{any additional parameters used by rgl::plot3d}
}
\value{
a 3d scatter plot
}
\description{
3d Plot
}
\examples{
data("swift")
start = as.POSIXct("2016-09-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-04-15","\%Y-\%m-\%d", tz="UTC")
swift = cutPAM(swift, start, end)
PAM_data = swift

calibration = triMAG(dta = PAM_data$magnetic)

plot3D(PAM_data$magnetic$mX, PAM_data$magnetic$mY, PAM_data$magnetic$mZ,
       xlab= "X", ylab= "Y", zlab= "Z")

}
