% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactive_timeseries_classification.R
\name{plot_interactive_timeseries_classification}
\alias{plot_interactive_timeseries_classification}
\title{Plot PAM data as an interactive timeseries}
\usage{
plot_interactive_timeseries_classification(
  to_classify,
  classification_datetime,
  classification,
  to_plot = c("light", "pressure", "temperature", "acceleration", "magnetic")
)
}
\arguments{
\item{to_classify}{rolling window data to plot}

\item{classification_datetime}{dates associated with classification}

\item{classification}{output of the classification}

\item{to_plot}{names of the variables to plot. For now this includes \code{light}, \code{pressure}, \code{acceleration} and \code{temperature}}
}
\value{
a plot of all the measurements
}
\description{
This opens a java dygraph application which allows the user to zoom in and out. In Rstudio it will open in the viewer pane and in base R in an html readers. Note that this can be a bit slow
}
\examples{
#load dummy data
data(hoopoe)
PAM_data=hoopoe

# This bit is for Rstudio users to prevent html from opening in Viewer pane and crashing
# It opens in web browser instead
backup_options <- options()
options(viewer=NULL)

# Plot interactive graphics
plot_interactive_timeseries(dta = PAM_data)

# restore Rstudio settings from before plot
options(backup_options)


}
\references{
Vanderkam, D., Allaire, J., Owen, J., Gromer, D., Shevtsov, P. and Thieurmel, B., dygraphs: Interface to Dygraphs Interactive Time Series Charting Library, 2015. URL http://CRAN. R-project. org/package= dygraphs. R package version 0.4, 5, p.7.
}
