% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensorimage_legend.R
\name{plot_sensorimage_legend}
\alias{plot_sensorimage_legend}
\title{sensor image legend}
\usage{
plot_sensorimage_legend(
  dta,
  col_palette,
  inset = c(0.5, -0.12),
  cex = 1,
  ncol = 5,
  bg = NA,
  ...
)
}
\arguments{
\item{dta}{data to be plotted (for instance temperature or pressure)}

\item{col_palette}{color palette to use in the graphic}

\item{inset}{where to place the legend default is at the bottom -0.12}

\item{cex}{size of labels}

\item{ncol}{number of columns in legend, default is 5}

\item{bg}{whether or not to have a background, default is NA}

\item{...}{Any additional parameters used by graphics::legend}
}
\value{
an image of the sensor data, for instance with activity it would produce an actogram
}
\description{
This function adds a legend to a sensor image
}
\examples{
#specify the data location
data(hoopoe)
start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")
PAM_data = create_crop(hoopoe,start,end)

# make margins big so that there is enough space for margin
par( mfrow= c(1,1), oma=c(4,2,0,6))

col_palette = c("black",viridis::cividis(90))
par(mar =  c(4,2,4,2))
plot_sensorimage(PAM_data$acceleration$date,
          PAM_data$acceleration$act, main = "Activity",
          col=col_palette, cex=1.2, cex.main = 2)
plot_sensorimage_legend(PAM_data$acceleration$act, col_palette)

}
