% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Quickly plot the pam data to have an idea of it's quality}
\usage{
plot_timeseries(
  dta,
  measurements = c("pressure", "light", "acceleration", "temperature", "magnetic"),
  ...
)
}
\arguments{
\item{dta}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be plotted. Currently supports these file extentions: "pressure","light", "acceleration", "temperature" and "magnetic"}

\item{...}{any additional parameters used by graphics::plot}
}
\value{
a plot of PAM data
}
\description{
Quickly plot the pam data to have an idea of it's quality
}
\examples{
PAM_data = hoopoe

#plot everything in 2 windows
par(mar=c(2.5,4,0.5,1))
plot_timeseries(PAM_data)

# only subset some measurements
plot_timeseries(PAM_data, measurements = c("light",
      "pressure", "acceleration"))

}
