% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotACTOGRAM.R
\name{plotACTOGRAM}
\alias{plotACTOGRAM}
\title{Plot Actogram}
\usage{
plotACTOGRAM(date, activity, tz = "UTC", offset = 0, dt = NA,
  xlab = "Hour", ylab = "Date", cex = 2, colours = c("black",
  viridis::magma(90)), par = par(mar = c(4, 4, 1, 6)), ...)
}
\arguments{
\item{date}{Date data in POSIXct format, most commonly `PAM_data$acceleration$date`}

\item{activity}{Activity data in POSIXct format, most commonly `PAM_data$acceleration$act`}

\item{tz}{Time zone for POSIXct, default set to "UTC"}

\item{offset}{This parameter determines where the center of the graph is. When `offset = 0`, then midday is at the center of the graph. when `offset=12` midnight`}

\item{dt}{the time interval to which the data are resampled (secs). Default is `NA`}

\item{xlab}{label for x-axis}

\item{ylab}{label for y axis}

\item{cex}{size of labels}

\item{par}{Graphical parameters. see `?par` for more details. default is `par = par(mar=c(4,4,1,6))`}

\item{type}{The type of classification used i.e "flapping" or soar-gliding}

\item{classification}{Is the result of the classification. is in the format of a vector of numbers used for low or high activity}

\item{threshold}{The threshold between different classes}

\item{date}{Date data in POSIXct format, most commonly `PAM_data$acceleration$date`}

\item{col}{Colour scheme of plot. Default `col = c("black",viridis::magma(90))`}
}
\value{
an actogram
}
\description{
Plot Actogram
}
\examples{
#specify the data location
data(PAM_data)

plotACTOGRAM <- function (date = PAM_data$acceleration$date,
                          activity = PAM_data$acceleration$act, tz="UTC",
                          offset = 0, ...)

}
