% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPLOT.R
\name{quickPLOT}
\alias{quickPLOT}
\title{quickly plot the pam data to have an idea of it's quality}
\usage{
quickPLOT(dta, measurements = c("pressure", "light", "acceleration",
  "temperature", "magnetic"), ...)
}
\arguments{
\item{dta}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be plotted. Currently supports these file extentions: "pressure","light", "acceleration", "temperature" and "magnetic"}

\item{...}{any additional parameters used by graphics::plot}
}
\value{
a plot of PAM data
}
\description{
quickly plot the pam data to have an idea of it's quality
}
\examples{
data(hoopoe)
PAM_data=hoopoe
str(PAM_data)

#plot everything in 2 windows
quickPLOT(PAM_data)

# only subset some measurements
quickPLOT(PAM_data, measurements = c("light", "pressure", "acceleration"))

}
