% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducePAM.R
\name{reducePAM}
\alias{reducePAM}
\title{Reduces the data to a specified temporal resolution}
\usage{
reducePAM(dta, varint = c("pressure", "light", "pit", "act",
  "temperature", "gX", "gY", "gZ", "mX", "mY", "mZ"))
}
\arguments{
\item{dta}{raw pam data see `data(bee_eater` for example}

\item{varint}{the variable of interest. For instance if all other variables should be summarised at the same temporal resolution as this specified variable. It must correspond to variable or list in `dta`}
}
\value{
reduced/summarised and interpolated dataset
}
\description{
Reduces the data to a specified temporal resolution
}
\examples{
data(bee_eater)
PAM_data = bee_eater
reduced_dta = reducePAM(PAM_data, interp = T)
head(reduced_dta)

}
