% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollPAM.R
\name{rollPAM}
\alias{rollPAM}
\title{Derive classification data for soaring birds}
\usage{
rollPAM(dta, interp_NA = TRUE, resolution_out = 15, window = 120,
  units = "mins")
}
\arguments{
\item{dta}{PAM data to be used in the analysis}

\item{interp_NA}{By default TRUE. Whether or not to interpolate NAs in dataset that the rollapply is used on}

\item{resolution_out}{Temporal resolution of output dataset. By defaukt 15 mins. Must be in minutes unless the units are changed}

\item{window}{Window over which to apply the rolling window. By defaukt 120 mins. Equivalent to zoo::rollapply(,width = window,) Must be in minutes unless the units are changed}

\item{units}{By default"mins", but also supports "hours" and "secs"}
}
\value{
a dataframe of derived metrics including the median, standard deviation, minimum, maximum, cumulative difference and range over the period
}
\description{
Derive classification data for soaring birds
}
\examples{
data(swift)
PAM_data = swift

# crop the data to get rid of no good periods
start = as.POSIXct("2016-09-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-04-21","\%Y-\%m-\%d", tz="UTC")
PAM_data = cutPAM(PAM_data, start, end)

TOclassify = rollPAM(dta = list(PAM_data$pressure, PAM_data$acceleration),
                     resolution_out = 15 ,
                     window = 120)



}
