% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOARprep.R
\name{SOARprep}
\alias{SOARprep}
\title{Derive classification data for soaring birds}
\usage{
SOARprep(dta, availavariable = c("pressure", "light", "acceleration"),
  diff_P = 2, twl, tz = "UTC")
}
\arguments{
\item{dta}{PAM data to be used in the analysis}

\item{availavariable}{Variables to be used to derive metrics for classification. must have "pressure", but ideally `availavariable = c("pressure", "light", "acceleration")` if any of these are incomplete, do not use them}

\item{diff_P}{pressure threshold. This if pressure changes more than e.g. 2hpa over 30 minutes, then the bird is flying}

\item{twl}{twilight estimates formatted according to twilightCalc in GeoLight package}

\item{tz}{Timeuzone. default is "UTC"}
}
\value{
a dataframe of derives metrhics based on pressure: `date`,`start` of flight event,`end`,`duration` of flight event,`total_daily_duration` spent flying that day,`total_daily_flight_number` total number of flight events that day,`pressure_change` absolute total pressure differences during that flight,`P_dep_arr` difference in pressure between the start of the flight and the end of the flight,`pressure_range` difference between min and max pressure during that flight,`mean_night_P` difference in mean pressure from the night before and the night after the flight event,`sd_night_P`,`mean_nextnight_P`,`sd_nextnight_P`,`night_P_diff`,`sum_activity` total amount of activity during that flight event,`prop_resting` proportion of time spent resting during that flight,`prop_active` proportion of time spent active during that flight event
}
\description{
Derive classification data for soaring birds
}
\examples{
data(PAM_data)
twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs, LightThreshold = 2, ask = F)
availavariable = c("pressure", "light", "acceleration")

TOclassify = SOARprep(dta = PAM_data, availavariable = c("pressure", "acceleration", "light"), twl = twl)

}
