% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sustainedACT.R
\name{sustainedACT}
\alias{sustainedACT}
\title{Make timetable}
\usage{
sustainedACT(dta, period = 3, toPLOT = T, act_threshold = 1,
  duration_threshold = 2, keep_one_off_missclassifications = FALSE,
  method = c("manual", "kmeans"), tz = "UTC")
}
\arguments{
\item{dta}{the input light data to be classified e.g. bee_eater$light}

\item{period}{number of timepoints after which behaviour is considered soaring e.g. for light recorded every 5 minutesand period = 3, then 3x5min = 15 minutes of continuous light is considered soaring flight}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{act_threshold}{This is the threshold between no activity and activity, be default it is set to 1}

\item{duration_threshold}{the duration in hours where flight is not longer soaring, but migratory}

\item{keep_one_off_missclassifications}{Can be TRUE or FALSE, if true, any one-off soaring missclassifications are replaced with soaring}

\item{method}{for the time being only supports "kmeans" and "manual", if methods= "kmeans" then duration_threshold is overwritten}

\item{tz}{timezone, default is "UTC"}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
data("bee_eater")

start = as.POSIXct("2015-07-12","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2016-06-01","\%Y-\%m-\%d", tz="UTC")

bee_eater = cutPAM(bee_eater, start, end)
quickPLOT(bee_eater, measurements = c("pressure", "light"))

dta = bee_eater$acceleration

behaviour = sustainedACT(dta , period = 3, toPLOT = T,
                         act_threshold = 1,
                         keep_one_off_missclassifications = FALSE,
                         method = "kmeans", tz= "UTC")

pressure_states = merge(bee_eater$pressure,
                        data.frame(bee_eater$acceleration,states=behaviour$classification ),
                        all=F)
par(mfrow=c(2,1), mar=c(4,4,0.5,0.5))
col=c("black","brown","cyan4","gold")
plot(pressure_states$date, pressure_states$obs, type="o",
     bg= col[pressure_states$states+1], pch=21,
     xlab="Date", ylab="Pressure",
     cex = ifelse(pressure_states$states==behaviour$migrating,1,0.7))

#migration is well classified, but soaring is over estimated
behaviour = sustainedACT(dta , period = 3, toPLOT = F,
                         act_threshold = 1,
                         duration_threshold=1.5,
                         keep_one_off_missclassifications = T,
                         method = "manual", tz= "UTC")

pressure_states = merge(bee_eater$pressure,
                        data.frame(bee_eater$acceleration,states=behaviour$classification ),
                        all=F)

col=c("black","brown","cyan4","gold")
plot(pressure_states$date, pressure_states$obs, type="o",
     bg= col[pressure_states$states+1], pch=21,
     xlab="Date", ylab="Pressure",
     cex = ifelse(pressure_states$states==behaviour$migrating,1,0.7))

}
