% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.R
\docType{data}
\name{swift}
\alias{swift}
\title{PAM data collected from Alpine Swift}
\format{
List of 6 variables
\describe{
  \item{id}{Logger id}
  \item{pressure}{Date in UTC and Atmospheric pressure measurements in hectopascals (every 15 minutes)}
  \item{light}{Date in UTC and Light intensity (every 2 minutes)}
  \item{acceleration}{Date in UTC, pitch of the accelerometer (i.e. an angle), and derived activity (every 5 minutes)}
  \item{temperature}{Date in UTC and temperature on bird in celcius(every 15 minutes)}
  \item{magnetic}{Date in UTC and magnetic measurements on 3 axes(every 4 hours)}
}
}
\source{
<www.vogelwarte.ch/en>
}
\usage{
swift
}
\description{
A dataset containing pressure, light, acceleration, temperature and magnetism from an alpine swift (id = 16FY)
}
\keyword{datasets}
