% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIFTprep.R
\name{SWIFTprep}
\alias{SWIFTprep}
\title{Derive classification data for soaring birds}
\usage{
SWIFTprep(dta, availavariable = c("pressure", "light", "acceleration"),
  twl, method = "manual", duration_threshold = 1,
  keep_one_off_missclassifications = FALSE, toPLOT = FALSE,
  tz = "UTC", ...)
}
\arguments{
\item{dta}{PAM data to be used in the analysis}

\item{availavariable}{Variables to be used to derive metrics for classification. must have "pressure", but ideally `availavariable = c("pressure", "light", "acceleration")` if any of these are incomplete, do not use them}

\item{twl}{twilight estimates formatted according to twilightCalc in GeoLight package}

\item{method}{for the time being only supports "kmeans" and "manual", if methods= "kmeans" then duration_threshold is overwritten}

\item{duration_threshold}{the duration in hours where flight is not longer soaring, but migratory}

\item{keep_one_off_missclassifications}{Can be TRUE or FALSE, if true, any one-off soaring missclassifications are replaced with soaring}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{tz}{Timeuzone. default is "UTC"}

\item{...}{any input for `sustainedACT()` function}
}
\value{
a dataframe of derives metrhics based on pressure: `date`,`start` of flight event,`end`,`duration` of flight event,`total_daily_duration` spent flying that day,`total_daily_flight_number` total number of flight events that day,`pressure_change` absolute total pressure differences during that flight,`P_dep_arr` difference in pressure between the start of the flight and the end of the flight,`pressure_range` difference between min and max pressure during that flight,`mean_night_P` difference in mean pressure from the night before and the night after the flight event,`sd_night_P`,`mean_nextnight_P`,`sd_nextnight_P`,`night_P_diff`,`sum_activity` total amount of activity during that flight event,`prop_resting` proportion of time spent resting during that flight,`prop_active` proportion of time spent active during that flight event
}
\description{
Derive classification data for soaring birds
}
\examples{
data(swift)
PAM_data = swift

# crop the data to get rid of no good periods
start = as.POSIXct("2016-09-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-04-21","\%Y-\%m-\%d", tz="UTC")
PAM_data = cutPAM(PAM_data, start, end)

twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs,
                             LightThreshold = 2, ask = FALSE)
availavariable = c("pressure", "light", "acceleration")

TOclassify = SWIFTprep(dta = PAM_data,
                      availavariable = c("pressure", "acceleration", "light"),
                      twl = twl)


}
