% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triACC.R
\name{triACC}
\alias{triACC}
\title{Calculate roll, pitch and yaw from tri axial acceleration data}
\usage{
triACC(dta)
}
\arguments{
\item{dta}{magentic data from PAM logger see hoopoe$magnetic for an example}
}
\value{
roll, pitch and yaw from acceleration data
}
\description{
Calculate roll, pitch and yaw from tri axial acceleration data
}
\examples{
data(swift)
PAM_data = swift

triACC(dta = PAM_data$magnetic)

}
