% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triMAG.R
\name{triMAG}
\alias{triMAG}
\title{Calibrate magnetic data and calculate roll, pitch and yaw}
\usage{
triMAG(dta)
}
\arguments{
\item{dta}{magentic data from PAM logger}
}
\value{
roll, pitch and yaw from acceleration data, as well as calibrated magnetic data for x, y and z axes
}
\description{
Calibrate magnetic data and calculate roll, pitch and yaw
}
\examples{
data(swift)
PAM_data = swift

calibration = triMAG(dta = PAM_data$magnetic)

}
