% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_crop.R
\name{wrangle_crop}
\alias{wrangle_crop}
\title{Crop all sensor data to the same timeframe}
\usage{
wrangle_crop(dta, start, end)
}
\arguments{
\item{dta}{path where files are stored}

\item{start}{posicxt object for date that PAM data should start}

\item{end}{posicxt object for date that PAM data should end}
}
\value{
shortened PAM data
}
\description{
Get rid of excess data. e.g. when a logger is kept in a rucksack or a lab before being downloaded.
}
\examples{
data(hoopoe)
PAM_data=hoopoe
str(PAM_data)

start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")

newPAM = wrangle_crop(PAM_data,start,end)

}
