% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_import.R
\name{wrangle_import}
\alias{wrangle_import}
\title{Import PAM data}
\usage{
wrangle_import(
  pathname = pathname,
  measurements = c(".pressure", ".glf", ".acceleration", ".temperature", ".magnetic")
)
}
\arguments{
\item{pathname}{path where files are stored}

\item{measurements}{a series of measurements logged by the PAM logger which are to be imported. Currently supports these file extentions: ".pressure", ".glf", ".gle",".acceleration", ".temperature", "AirTemperature", ".BodyTemperature" and ".magnetic"}
}
\value{
a list of measurements for the one individual
}
\description{
Imports and formats many datasets into one big nested list containing all the data from the different sensors. A subset of sensors can be selected using \code{measurements}.
}
\examples{
#pathname = "your/path/here"
#measurements = c(".pressure", ".glf")
#PAM_data = wrangle_import(pathname, measurements)
#str(PAM_data)
#plot(PAM_data$light$date[3000:5000], PAM_data$light$obs[3000:5000],
#type="l", xlab="Date", ylab="Light Intensity")

}
