% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitudeCALC.R
\name{altitudeCALC}
\alias{altitudeCALC}
\title{Calculate altitude from pressure}
\usage{
altitudeCALC(P, T0 = 288.15, P0 = 1013.25)
}
\arguments{
\item{P}{on-bird pressure in hectopascals}

\item{T0}{temperature at sea level in kelvin (i.e. celcius + 273.15 OR (Farenheit − 32) * 5/9 + 273.15)}

\item{P0}{pressure at sea level in hectopascals}
}
\value{
altitude in metres, default is calculated according to International Standard Atmosphere model (International Organization for Standardization 1975: ISO 2533:1975)
}
\description{
This function calculated altitude from atmospheric pressure
}
\examples{
altitude = altitudeCALC(P = hoopoe$pressure$obs)
plot(hoopoe$pressure$date,
     altitude,
     type="o",
     pch=16,
     xlab="Date",
     ylab="Altitude (m)")

}
\references{
Atmosphere, S., 1975. International organization for standardization. ISO, 2533, p.1975.
}
