% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification2PAM.R
\name{classification2PAM}
\alias{classification2PAM}
\title{Integrate classification from classifyPAM into PAM data}
\usage{
classification2PAM(from, to, classification, addTO)
}
\arguments{
\item{from}{start of event that was classified (generally SOARprep output)}

\item{to}{end of event that was classified (generally SOARprep output)}

\item{classification}{of event (generally classifyPAM()$states output )}

\item{addTO}{data which the classifications are to be added to (e.g. PAM_data$pressure)}
}
\value{
the classification in addTO dataset
}
\description{
Integrate classification from classifyPAM into PAM data
}
\examples{
data(bee_eater)
PAM_data = bee_eater
twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs,
LightThreshold = 2, ask = F)
availavariable = c("pressure", "light", "acceleration")

TOclassify = SOARprep(dta = PAM_data, availavariable = availavariable, twl = twl,diff_P=2)

classification = classifyPAM(TOclassify$total_daily_duration * log(TOclassify$night_P_diff+0.001 )
* TOclassify$total_daily_P_change,
states=3, "hmm")$state
pressure_classification = classification2PAM(from = TOclassify$start,
to =TOclassify$end,
classification = classification,
addTO = PAM_data$pressure)

plot(PAM_data$pressure$date, PAM_data$pressure$obs,
col= viridis::viridis(4)[pressure_classification+1],
type="o", pch=16, cex=0.6)


}
