% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyPAM.R
\name{classifyPAM}
\alias{classifyPAM}
\title{Derive classification data for soaring birds}
\usage{
classifyPAM(dta, method = c("kmeans", "hmm"), states = 2, ...)
}
\arguments{
\item{dta}{data to be classified, can be anything}

\item{method}{method for classifying data, currently support "hmm" for hidden markov model and "kmeans" for kmeans clustering}

\item{states}{number of states to classify the data into}

\item{...}{any additional inputs for depmixs4::depmix or or stats::kmeans functions, depending on method selected}
}
\value{
the data's classification based on the chosen algorithm
}
\description{
Derive classification data for soaring birds
}
\examples{
data(hoopoe)
PAM_data=hoopoe

twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs,
LightThreshold = 2, ask = F)
availavariable = c("pressure", "light", "acceleration")

TOclassify = SOARprep(dta = PAM_data, availavariable = availavariable, twl = twl)

classification = classifyPAM(TOclassify$night_P_diff, states=3, "hmm")$state
pressure_classification = classification2PAM(from = TOclassify$start,
to =TOclassify$end,
classification = classification,
addTO = PAM_data$pressure)

plot(PAM_data$pressure$date, PAM_data$pressure$date, col= pressure_classification+1, type="o")

}
