% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySOAR.R
\name{classifySOAR}
\alias{classifySOAR}
\title{Make timetable}
\usage{
classifySOAR(dta, toPLOT = T, method = "manual", threshold = 2,
  period = 2)
}
\arguments{
\item{dta}{data stored as a list see str(data(PAM_data)) for example format}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{method}{for the time being only supports manual. In this case the threshold = 2 as a default. This is the 15 minute hectopascal change}

\item{threshold}{the manual threshold at which activity and non activity is drawn. in this case a change of greater than 2 hpa /15 minutes ios unlikely to have be caused by weather}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 1x15min = 15 minutes of intense activity is considered a migratory flight}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
#specify the data location
data(PAM_data)
str(PAM_data)

#plot the activity to see if it looks ok
plot(PAM_data$pressure$date, PAM_data$pressure$obs, xlab = "Time", ylab = "Pressure (hPa)")

# at first glance it looks like the logger was removed off a birds and left in arucksack
PAM_data$pressure = PAM_data$pressure[((PAM_data$pressure$date >= "2016-07-30")
& (PAM_data$pressure$date <= "2017-06-01")),]

behaviour = classifySOAR(dta = PAM_data$pressure, period = 2, toPLOT = F)


col=col=c("brown","cyan4","black","gold")
plot(PAM_data$pressure$date[2700:3000], PAM_data$pressure$obs[2700:3000],
col=col[behaviour$classification][2700:3000], type="o", pch=20, xlab="Date", ylab="Pressure")

behaviour$timetable

}
