% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifySOAR.R
\name{classifySOAR}
\alias{classifySOAR}
\title{Make timetable}
\usage{
classifySOAR(dta, toPLOT = TRUE, method = "PressureChange",
  threshold = 2, period = 2, tz = "UTC")
}
\arguments{
\item{dta}{data stored as a list see str(data(hoopoe)) for example format}

\item{toPLOT}{can be true or false. If true then threshold is plotted according to plotTHLD()}

\item{method}{for the time being only supports manual. In this case the threshold = 2 as a default. This is the 15 minute hectopascal change}

\item{threshold}{the manual threshold at which activity and non activity is drawn. in this case a change of greater than 2 hpa /15 minutes ios unlikely to have be caused by weather}

\item{period}{number of timepoints after which behaviour is considered migratory e.g. for hoopoes, 1x15min = 15 minutes of intense activity is considered a migratory flight}

\item{tz}{timezone, default is "UTC"}
}
\value{
a timetable for when the species was migrating or not
}
\description{
Make timetable
}
\examples{
#specify the data location
data(hoopoe)
str(hoopoe)

#plot the activity to see if it looks ok
plot(hoopoe$pressure$date, hoopoe$pressure$obs, xlab = "Time", ylab = "Pressure (hPa)")

# at first glance it looks like the logger was removed off a birds and left in arucksack
hoopoe$pressure = hoopoe$pressure[((hoopoe$pressure$date >= "2016-07-30")
& (hoopoe$pressure$date <= "2017-06-01")),]

behaviour = classifySOAR(dta = hoopoe$pressure, period = 2, toPLOT = FALSE)

#plot the rclassification as an interactive plot
backup_options <- options()
options(viewer=NULL) # ensure it is viewed in internet browser
dygraphClassified(dta = hoopoe, timetable = behaviour$timetable)
options(backup_options) # restore previous viewer settings

col=col=c("brown","cyan4","black","gold")
plot(hoopoe$pressure$date[2700:7000], hoopoe$pressure$obs[2700:7000],
col=col[behaviour$classification][2700:7000], type="o", pch=20, xlab="Date", ylab="Pressure")

behaviour$timetable
}
