% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutPAM.R
\name{cutPAM}
\alias{cutPAM}
\title{Crop all sensor data to the same time}
\usage{
cutPAM(dta, start, end)
}
\arguments{
\item{dta}{path where files are stored}

\item{start}{posicxt object for date that PAM data should start}

\item{end}{posicxt object for date that PAM data should end}
}
\value{
shortened PAM data
}
\description{
Get rid of excess data. e.g. when a logger is kept in a rucksack or a lab before being downloaded.
}
\examples{
#data(hoopoe)
#PAM_data=hoopoe
#str(PAM_data)

#start = as.POSIXct("2016-07-01","\%Y-\%m-\%d", tz="UTC")
#end = as.POSIXct("2017-06-01","\%Y-\%m-\%d", tz="UTC")

#newPAM = cutPAM(PAM_data,start,end)

}
