% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraphPAM.R
\name{dygraphPAM}
\alias{dygraphPAM}
\title{Plot PAM data with dygraphs}
\usage{
dygraphPAM(
  dta,
  from = dta$light$date[1],
  to = dta$light$date[length(dta$light$date)],
  toPLOT = names(dta)
)
}
\arguments{
\item{dta}{PAM data to be plotted}

\item{from}{date that plotting starts}

\item{to}{date that plotting ends}

\item{toPLOT}{names of the variables to plot. For now this includes `light`, `pressure`, `acceleration` and `temperature`}
}
\value{
a plot of all the measurements
}
\description{
This opens a java dygraph application which allows the user to zoom in and out. In Rstudio it will open in the viewer pane and in base R in an html readers. Note that this can be a bit slow
}
\examples{
#load dummy data
data(hoopoe)
PAM_data=hoopoe

# This bit is for Rstudio users to prevent html from opening in Viewer pane and crashing
# It opens in web browser instead
backup_options <- options()
options(viewer=NULL)

# Plot interactive graphics
dygraphPAM(dta = PAM_data)

# restore Rstudio settings from before plot
options(backup_options)


}
\references{
Vanderkam, D., Allaire, J., Owen, J., Gromer, D., Shevtsov, P. and Thieurmel, B., dygraphs: Interface to Dygraphs Interactive Time Series Charting Library, 2015. URL http://CRAN. R-project. org/package= dygraphs. R package version 0.4, 5, p.7.
}
