% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePAM.R
\name{mergePAM}
\alias{mergePAM}
\title{Integrate classification from classifyPAM into PAM data}
\usage{
mergePAM(PAM_data, interp = T, keep_NA = T)
}
\arguments{
\item{PAM_data}{raw pam data see `data(bee_eater` for example}

\item{interp}{logical. whether or not to interpolate if there are NAs}

\item{keep_NA}{logical. Whether or not to keep NA}
}
\value{
merged and interpolated dataset
}
\description{
Integrate classification from classifyPAM into PAM data
}
\examples{
data(bee_eater)
PAM_data = bee_eater
merged_dta = mergePAM(PAM_data, interp = T)
head(merged_dta)

}
