% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pamPREP.R
\name{pamPREP}
\alias{pamPREP}
\title{Derive classification data for soaring birds}
\usage{
pamPREP(PAM_data, availavariable = c("pressure", "light", "acceleration",
  "magnetic", "temperature"), Pdiff_thld = 2, light_thld = 1,
  method = "pressure", twl, interp = TRUE, tz = "UTC")
}
\arguments{
\item{PAM_data}{PAM data to be used in the analysis}

\item{availavariable}{Variables to be used to derive metrics for classification. must have "pressure", but ideally `availavariable = c("pressure", "light", "acceleration")` if any of these are incomplete, do not use them}

\item{Pdiff_thld}{Pressure threshold. Only used when method="pressure".  This if pressure changes more than e.g. 2hpa over 30 minutes, then the bird is flying}

\item{light_thld}{Light threshold. Only used when method="darkness". This is the the light threshold for finding darkness, should be the same as for GeoLight::twilightCalc}

\item{method}{The type of event that is being classified. Can be "flap", "endurance", "rest", "pressure","light" or "darkness".If method = "pressure" then  it find periods where pressure has changed more than a certain threshold. If method = "flap", then the algorithm looks for sustained periods of high activity. If method = "endurance" it looks for sustained activity (low or high). If method = "rest+ then it looks for sustained periods of no activity. If method = "light" if looks for periods of sustained sunlight. If method = "darkness" if looks for periods of darkness}

\item{twl}{twilight estimates formatted according to twilightCalc in GeoLight package}

\item{interp}{whether or not to interpolate the magnetic data. If FALSE, then NAs are left in the dataset}

\item{tz}{Timeuzone. default is "UTC"}
}
\value{
a dataframe of derives metrhics based on pressure: `date`,`start` of event event,`end`,`duration` of event event,`total_daily_duration` spent flying that day,`total_daily_event_number` total number of event events that day,`pressure_change` absolute total pressure differences during that event,`P_dep_arr` difference in pressure between the start of the event and the end of the event,`pressure_range` difference between min and max pressure during that event,`mean_night_P` difference in mean pressure from the night before and the night after the event event,`sd_night_P`,`mean_nextnight_P`,`sd_nextnight_P`,`night_P_diff`,`sum_activity` total amount of activity during that event event,`prop_resting` proportion of time spent resting during that event,`prop_active` proportion of time spent active during that event event
}
\description{
Derive classification data for soaring birds
}
\examples{
data(hoopoe)
PAM_data=hoopoe
twl = GeoLight::twilightCalc(PAM_data$light$date, PAM_data$light$obs,
                             LightThreshold = 2, ask = FALSE)

TOclassify = pamPREP(PAM_data = PAM_data,
                      method= "flap",
                      twl = twl)


}
