% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTHLD.R
\name{plotTHLD}
\alias{plotTHLD}
\title{Plot threshold}
\usage{
plotTHLD(dta, type = "flapping", classification, threshold)
}
\arguments{
\item{dta}{Raw acceleration or pressure data used to make the classification}

\item{type}{The type of classification used i.e "flapping" or soar-gliding}

\item{classification}{Is the result of the classification. is in the format of a vector of numbers used for low or high activity}

\item{threshold}{The threshold between different classes}
}
\value{
a graphic with the output from the classifications
}
\description{
Plot threshold
}
\examples{
#specify the data location
data(PAM_data)

PAM_data$acceleration = PAM_data$acceleration[((PAM_data$acceleration$date >= "2016-07-30")
& (PAM_data$acceleration$date <= "2017-06-01")),]

behaviour = classifyFLAP(dta = PAM_data$acceleration, flapping_duration = 3, toPLOT = F)

plotTHLD(dta = PAM_data$acceleration$act,
     type = behaviour$type,
     classification = behaviour$classification,
     threshold = behaviour$threshold)

}
