% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensorIMG.R
\name{sensorIMG}
\alias{sensorIMG}
\title{Plot Actogram}
\usage{
sensorIMG(date, sensor_data, tz = "UTC", plotx = TRUE, ploty = TRUE,
  offset = 0, dt = NA, xlab = "Hour", ylab = "Date", cex = 2,
  col = c("black", viridis::magma(90)), ...)
}
\arguments{
\item{date}{Date data in POSIXct format, most commonly `PAM_data$acceleration$date`}

\item{sensor_data}{sensor data, for example look at `PAM_data$acceleration$act`}

\item{tz}{Time zone for POSIXct, default set to "UTC"}

\item{plotx}{wherether or not to plot the x axis (for instance when compiling multifigures)}

\item{ploty}{wherether or not to plot the y axis (for instance when compiling multifigures)}

\item{offset}{This parameter determines where the center of the graph is. When `offset = 0`, then midday is at the center of the graph. when `offset=12` midnight`}

\item{dt}{the time interval to which the data are resampled (secs). Default is `NA`}

\item{xlab}{label for x-axis}

\item{ylab}{label for y axis}

\item{cex}{size of labels}

\item{col}{Colour scheme of plot. Default `col = c("black",viridis::magma(90))`}

\item{...}{Any additional parameters used by graphics::image}
}
\value{
an actogram
}
\description{
Plot Actogram
}
\examples{
#specify the data location
data(hoopoe)
PAM_data=hoopoe

# Create plots with large margins (mar) and plot 3 together (mfrow)
par(mar = c(4,4,1,6),  mfrow= c(1,3))

sensorIMG(PAM_data$acceleration$date,
          PAM_data$acceleration$act, main = "Activity")

sensorIMG(PAM_data$pressure$date,
          PAM_data$pressure$obs,  main="Pressure")

sensorIMG(PAM_data$temperature$date,
          PAM_data$temperature$obs,  main="Temperature")

}
